requirements/base.txt
Django==1.5.1
ipython
django-braces==1.2.2
django-model-utils==1.4.0
logutils==0.3.3
South==0.8.1
django-extensions==1.2.2
Werkzeug==0.9.4
django-social-auth==0.7.28
facebook-sdk==0.4.0
gitpython
django-redis==3.3
django-celery==3.0.23
Pillow==2.1.0
django_ace
requests
pylint
